+++ apps/herdr-plugin-red-skills/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @reddb-io/herdr-plugin-red-skills
 
+## 3.8.0
+
+### Patch Changes
+
+- @reddb-io/build-info@3.8.0
+
+## 3.7.1
+
+### Patch Changes
+
+- @reddb-io/build-info@3.7.1
+
+## 3.7.0
+
+### Patch Changes
+
+- @reddb-io/build-info@3.7.0
+
 ## 3.6.2
 
 ### Patch Changes