+++ apps/vscode-extension-red-skills/CHANGELOG.md
@@ -1,5 +1,32 @@
 # vscode-extension-red-skills
 
+## 3.8.0
+
+### Patch Changes
+
+- @reddb-io/shared@3.8.0
+- @reddb-io/brand-tokens@3.8.0
+- @reddb-io/redskilled@3.8.0
+- @reddb-io/redskilled-render@3.8.0
+
+## 3.7.1
+
+### Patch Changes
+
+- @reddb-io/shared@3.7.1
+- @reddb-io/brand-tokens@3.7.1
+- @reddb-io/redskilled@3.7.1
+- @reddb-io/redskilled-render@3.7.1
+
+## 3.7.0
+
+### Patch Changes
+
+- @reddb-io/shared@3.7.0
+- @reddb-io/brand-tokens@3.7.0
+- @reddb-io/redskilled@3.7.0
+- @reddb-io/redskilled-render@3.7.0
+
 ## 3.6.2
 
 ### Patch Changes