+++ packages/github/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @reddb-io/github
 
+## 3.8.0
+
+## 3.7.1
+
+## 3.7.0
+
 ## 3.6.2
 
 ### Patch Changes