+++ packages/redskilled-render/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @reddb-io/redskilled-render
 
+## 3.8.0
+
+### Patch Changes
+
+- @reddb-io/shared@3.8.0
+
+## 3.7.1
+
+### Patch Changes
+
+- @reddb-io/shared@3.7.1
+
+## 3.7.0
+
+### Patch Changes
+
+- @reddb-io/shared@3.7.0
+
 ## 3.6.2
 
 ### Patch Changes